# api/api.yaml
openapi: 3.0.0
info:
  title: Sample API
  description: API for RecipeProvider
  version: 0.1.9
servers:
  - url: http://localhost:8080
    description: Optional server description, e.g. Main (production) server

paths:
  /recipes:
    get:
      summary: Return all recipes.
      description: Return all recipes.
      operationId: getAllRecipes
      responses:
        '200': # status code
          description: Return all recipe
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecipeReadDto'
    post:
      summary: Create recipes.
      description: Create recipes.
      operationId: createRecipes
      requestBody:
        required: true
        description: Recipes in JSON array format.
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/RecipeWriteDto'
      responses:
        '200': # status code
          description: Return the created recipes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecipeReadDto'
  /recipes/{id}:
    get:
      summary: Return a specific recipe by id.
      description: Return a specific recipe by id.
      operationId: getRecipeById
      parameters:
        - in: path
          name: id
          required: true
          description: The Id of the recipe
          schema:
            type: string
      responses:
        '200': # status code
          description: Return a specific recipe by id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeReadDto'
    delete:
      summary: Delete a specific recipe by id.
      description: Delete a specific recipe by id.
      operationId: deleteRecipeById
      parameters:
        - in: path
          name: id
          required: true
          description: The Id of the recipe
          schema:
            type: string
      responses:
        '200': # status code
          description: Return the deleted recipes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeReadDto'
  /recipes/menu:
    get:
      summary: Return the actual menu
      description: Return the actual menu
      operationId: getActualMenu
      parameters:
        - in: query
          name: random
          schema:
            type: boolean
            default: false
          description: Get a random menu or the actual menu
      responses:
        '200': # status code
          description: Return the actual menu
          content:
            application/json:
              schema:
                type: array
                minItems: 2
                maxItems: 2
                items:
                  $ref: '#/components/schemas/RecipeReadDto'
    put:
      summary: Update the actual menu
      description:  Update the actual menu
      operationId: updateMenu
      requestBody:
        required: true
        description: Recipes for the new menu
        content:
          application/json:
            schema:
              type: array
              minItems: 2
              maxItems: 2
              items:
                $ref: '#/components/schemas/RecipeReadDto'
      responses:
        '200': # status code
          description: Return the actual menu
          content:
            application/json:
              schema:
                type: array
                minItems: 2
                maxItems: 2
                items:
                  $ref: '#/components/schemas/RecipeReadDto'
components:
  schemas:
    RecipeReadDto:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        subtitle:
          type: string
        labels:
          type: array
          items:
            type: string
            enum:
              - Kalorienarm
              - Fettarm
              - Eiweiß
              - Vegetarisch
              - Vegan
              - Schwein
              - Rind
              - Geflügel
              - Fisch
              - Thermomix
              - Airfryer
              - Express
        image:
          type: string
        duration:
          type: integer
          description: Duration in minutes
          example: 30
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/AmountDto'
        nutrients:
          type: array
          items:
            $ref: '#/components/schemas/AmountDto'
        preparation:
          type: array
          items:
            type: string
      required:
        - id
        - title
    RecipeWriteDto:
      type: object
      properties:
        title:
          type: string
        subtitle:
          type: string
        labels:
          type: array
          items:
            type: string
            enum:
              - Kalorienarm
              - Fettarm
              - Eiweiß
              - Vegetarisch
              - Vegan
              - Schwein
              - Rind
              - Geflügel
              - Fisch
              - Thermomix
              - Airfryer
              - Express
        image:
          type: string
        duration:
          type: integer
          description: Duration in minutes
          example: 30
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/AmountDto'
        nutrients:
          type: array
          items:
            $ref: '#/components/schemas/AmountDto'
        preparation:
          type: array
          items:
            type: string
      required:
        - title
    AmountDto:
      type: object
      properties:
        name:
          type: string
        amount:
          type: integer
        unit:
          type: string
          enum:
            - g
            - kg
            - St
            - ml
            - l
            - Esslöfel
            - Teelöfel
            - kcal